+++ README.rst
@@ -29,8 +29,10 @@ all necessary requirements as follows::
   (vivarium_census_prl_synth_pop) :~$ git clone https://github.com/ihmeuw/vivarium_census_prl_synth_pop.git
   ...git will copy the repository from github and place it in your current directory...
   (vivarium_census_prl_synth_pop) :~$ cd vivarium_census_prl_synth_pop
-  (vivarium_census_prl_synth_pop) :~$ pip install -e .
+  (vivarium_census_prl_synth_pop) :~$ pip install -e .[dev]
   ...pip will install vivarium and other requirements...
+  (vivarium_census_prl_synth_pop) :~$ jobmon_config update --web_service_fqdn jobmon.scicomp.ihme.washington.edu --web_service_port 80
+  ...configures jobmon...
 
 
 Note the ``-e`` flag that follows pip install. This will install the python
@@ -115,3 +117,28 @@ The ``-v`` flag will log verbosely, so you will get log messages every time
 step. For more ways to run simulations, see the tutorials at
 https://vivarium.readthedocs.io/en/latest/tutorials/running_a_simulation/index.html
 and https://vivarium.readthedocs.io/en/latest/tutorials/exploration.html
+
+
+Running post-processing
+-----------------------
+
+Once the simulation is finished, post-processing can be run on the raw results by, e.g.::
+
+   (vivarium_census_prl_synth_pop) :~$ make_results <OUTPUT-DIR> <OPTIONS>
+
+The ``-j``, ``--jobmon`` flag uses jobmon to launch the post-processing 
+pipeline. If it is not included, then it will launch the pipeline on the current
+node.
+
+Note: if when running with ``--jobmon`` causes a No route to host error, it is
+possible that jobmon did not get installed correctly. In this event,
+
+   (vivarium_census_prl_synth_pop) :~$ pip install pip install jobmon_installer_ihme==<version>
+   (vivarium_census_prl_synth_pop) :~$ <jobmon-config-command>
+
+where <version> is the jobmon installer version pinned in the setup.py and
+<jobmon-config-command> is the corresponding command to run from
+https://hub.ihme.washington.edu/pages/viewpage.action?spaceKey=DataScience&title=Jobmon+Conda+Versions
+
+
+Jobmon docuemtation can be found at https://scicomp-docs.ihme.washington.edu/jobmon/current/
