+++ src/vivarium_census_prl_synth_pop/tools/__init__.py
@@ -1,3 +1,3 @@
 from .app_logging import configure_logging_to_terminal
 from .make_artifacts import build_artifacts
-from .make_results import build_results
+from .make_results import do_build_results
