+++ src/vivarium_census_prl_synth_pop/tools/cli.py
@@ -1,4 +1,5 @@
-from typing import Tuple
+from pathlib import Path
+from typing import Tuple, Union
 
 import click
 from loguru import logger
@@ -10,7 +11,7 @@ from vivarium_census_prl_synth_pop.tools import (
     configure_logging_to_terminal,
 )
 from vivarium_census_prl_synth_pop.tools.jobmon import run_make_results_workflow
-from vivarium_census_prl_synth_pop.tools.make_results import build_results
+from vivarium_census_prl_synth_pop.tools.make_results import do_build_results
 from vivarium_census_prl_synth_pop.utilities import build_final_results_directory
 
 
@@ -180,7 +181,42 @@ def make_results(
         func = run_make_results_workflow
         kwargs = cluster_requests
     else:  # run from current node
-        func = build_results
+        func = do_build_results
         kwargs = {}
     main = handle_exceptions(func=func, logger=logger, with_debugger=with_debugger)
     main(raw_output_dir, final_output_dir, mark_best, test_run, artifact_path, **kwargs)
+
+
+@click.command()
+@click.argument("raw_output_dir", type=click.Path(exists=True))
+@click.argument("final_output_dir", type=click.Path(exists=True))
+@click.option("-v", "verbose", count=True, help="Configure logging verbosity.")
+@click.option(
+    "--pdb",
+    "with_debugger",
+    is_flag=True,
+    help="Drop into python debugger if an error occurs.",
+)
+@click.option("-b", "--mark-best", is_flag=True)
+@click.option(
+    "-t",
+    "--test-run",
+    is_flag=True,
+)
+@click.option(
+    "-i", "--artifact-path", type=click.Path(exists=True), default=paths.DEFAULT_ARTIFACT
+)
+def build_results(
+    raw_output_dir: Path,
+    final_output_dir: Path,
+    verbose: int,
+    with_debugger: bool,
+    mark_best: bool,
+    test_run: bool,
+    artifact_path: Union[str, Path],
+) -> None:
+    configure_logging_to_terminal(verbose)
+    main = handle_exceptions(
+        func=do_build_results, logger=logger, with_debugger=with_debugger
+    )
+    main(raw_output_dir, final_output_dir, mark_best, test_run, artifact_path)
