+++ src/vivarium_census_prl_synth_pop/tools/jobmon.py
@@ -49,7 +49,7 @@ def run_make_results_workflow(
         },
         default_cluster_name="slurm",
         command_template=(
-            f"{shutil.which('make_results')} "
+            f"{shutil.which('build_results')} "
             "{raw_output_dir} "
             "{final_output_dir} "
             "{mark_best} "
